separating views.py
