+++ accounts/urls.py
@@ -1,11 +1,12 @@
 # from django.contrib.auth import views as auth_views
 from django.urls import path
-from . import views
-
+from .views import login_views,signup_views
 app_name = "accounts"
 urlpatterns = [
-    path("signup/", views.signup, name="signup"),
-    path('login/', views.login, name='login'),
+    # signup_views
+    path("signup/", signup_views.signup, name="signup"),
+    # login_views
+    path('login/', login_views.login, name='login'),
     # path('login/', auth_views.LoginView.as_view(template_name="accounts/login.html"), name='login'),
-    path('logout/', views.logout, name='logout'),
+    path('logout/', login_views.logout, name='logout'),
 ]
