+++ accounts/views/login_views.py
@@ -1,5 +1,5 @@
 from django.shortcuts import render, redirect
-from .forms import CustomUserCreationForm
+from ..forms import CustomUserCreationForm
 from django.contrib.auth import login as auth_login
 from django.contrib.auth import logout as auth_logout
 from django.contrib.auth.forms import UserCreationForm, AuthenticationForm
@@ -7,17 +7,6 @@ from django.contrib.auth.forms import UserCreationForm, AuthenticationForm
 # Create your views here.
 
 
-def signup(request):
-    if request.method == "POST":
-        form = CustomUserCreationForm(request.POST)
-        if form.is_valid():
-            user = form.save()
-            auth_login(request, user)
-            return redirect("reviews:index")
-    else:
-        form = CustomUserCreationForm()
-    context = {"form": form}
-    return render(request, "accounts/signup.html", context)
 
 
 def login(request):
@@ -34,6 +23,7 @@ def login(request):
     }
     
     return render(request, 'accounts/login.html', context)
+
 
 def logout(request):
     auth_logout(request)
