+++ accounts/views/signup_views.py
@@ -1,0 +1,20 @@
+from django.shortcuts import render, redirect
+from ..forms import CustomUserCreationForm
+from django.contrib.auth import login as auth_login
+from django.contrib.auth import logout as auth_logout
+from django.contrib.auth.forms import UserCreationForm, AuthenticationForm
+
+
+
+def signup(request):
+    if request.method == "POST":
+        form = CustomUserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save()
+            auth_login(request, user)
+            return redirect("reviews:index")
+    else:
+        form = CustomUserCreationForm()
+    context = {"form": form}
+    return render(request, "accounts/signup.html", context)
+
